requirements.txt
streamlit==1.55.0
tensorflow-cpu==2.16.1
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
protobuf==4.25.3
